added activity list
